move ParseExpression from parser into util/parser.
